server-java-modern: lint cleanup batch 2 (~50 warnings fixed)
Five categories of compiler warnings resolved:

* 17 [static] StringEncryption qualified-by-instance:
  - StringEncryption.cipherDictionary accessed via `this.` from
    inside instance methods (16 sites) — switched to bare reference
    so it's clear the field is static.
  - PacketBuilder routed through `DataConversions.encryption.encipher()`
    (instance ref to a static method); now calls `StringEncryption.encipher`
    directly + adds the import.

* 22 [cast] redundant int casts in plugins:
  - DataConversions.timeSince and time-since helpers (5 sites)
  - Formulae.getHeight: `(int) (y / 944)` where y is already int
  - BCrypt.crypt_raw: `(int[]) bf_crypt_ciphertext.clone()` — array
    .clone() is covariant, no cast needed
  - Underground Pass quest plugins (~20 sites of
    `(int) (getCurrentLevel(...) / N)` where the inner is int math)
  - Certer.certAmount, AgilityShortcuts (× 2), Thieving.incExp

* 3 [serial] missing serialVersionUID:
  - SpellFailureException, PluginInterruptedException,
    ScriptEndedException, AStarPathfinder.JPanel2

* 1 [unchecked] DiscordService auction list clone:
  - `(ArrayList<MarketItem>) auctions.clone()` replaced with
    `new ArrayList<>(auctions)` — safer + no unchecked cast.

Verified: ant compile clean, smoke_test.sh 4/4 pass.

+++ server-java-modern/plugins/com/openrsc/server/plugins/authentic/npcs/Certer.java
@@ -285,7 +285,7 @@ public class Certer implements TalkNpcTrigger, UseNpcTrigger {
 				player.message("As an Ultimate Ironman. you cannot use certer bank exchange.");
 				return;
 			}
-			certAmount = (int) (player.getBank().countId(itemID) / 5);
+			certAmount = player.getBank().countId(itemID) / 5;
 			int itemAmount = certAmount * 5;
 			if (itemAmount <= 0) {
 				player.message("You don't have any " + names[useIndex] + " to certificate");
